OWLS-91213: Shutdown servers using REST operations prior to deleting pods (#2635)
* Initial changes

* remove unused imports

* Add logging message and unit tests

* revert blank space

* Cleanup unused unit test code

* Support force shutdown of managed server's via REST

* documentation update

* Update number one based on review feedback

* Added retry on thrown exception

* Fix unit tests from merge and update copyright

* Add fudge factor to HTTP shutdown request

* Code cleanup based on review

* Documentation fixes

+++ operator/src/main/java/oracle/kubernetes/operator/steps/HttpRequestProcessing.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2021, Oracle and/or its affiliates.
+// Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 package oracle.kubernetes.operator.steps;
@@ -23,7 +23,7 @@ import oracle.kubernetes.operator.work.Packet;
 
 abstract class HttpRequestProcessing {
 
-  public static final Integer HTTP_TIMEOUT_SECONDS = 60;
+  public static final Long HTTP_TIMEOUT_SECONDS = 60L;
   protected final Packet packet;
   private final V1Service service;
   protected final V1Pod pod;
@@ -36,7 +36,7 @@ abstract class HttpRequestProcessing {
 
   static HttpAsyncRequestStep createRequestStep(HttpRequest request, HttpResponseStep responseStep) {
     return HttpAsyncRequestStep.create(request, responseStep)
-          .withTimeoutSeconds(HTTP_TIMEOUT_SECONDS);
+        .withTimeoutSeconds(HTTP_TIMEOUT_SECONDS);
   }
 
   @Nonnull
